comment update proccess

+++ src/modules/forum/useCases/comments/deleteComment/DeleteComment.ts
@@ -7,6 +7,7 @@ import { DeleteCommentDTO } from "./DeleteCommentDTO";
 import { DeleteCommentErrors } from "./DeleteCommentErrors";
 import { ICommentRepo } from "../../../repos/commentRepo";
 import { IMemberRepo } from "../../../repos/memberRepo";
+import { IPostRepo } from "../../../repos/postRepo";
 import { UniqueEntityID } from "../../../../../shared/domain/UniqueEntityID";
 import { UseCase } from "../../../../../shared/core/UseCase";
 
@@ -21,12 +22,18 @@ type Response = Either<
 export class DeleteComment
   implements UseCase<DeleteCommentDTO, Promise<Response>>
 {
+  private postRepo: IPostRepo;
   private commentRepo: ICommentRepo;
   private memberRepo: IMemberRepo;
 
-  constructor(commentRepo: ICommentRepo, memberRepo: IMemberRepo) {
+  constructor(
+    commentRepo: ICommentRepo,
+    memberRepo: IMemberRepo,
+    postRepo: IPostRepo
+  ) {
     this.commentRepo = commentRepo;
     this.memberRepo = memberRepo;
+    this.postRepo = postRepo;
   }
 
   public async execute(req: DeleteCommentDTO): Promise<Response> {
@@ -48,6 +55,13 @@ export class DeleteComment
       const commentIdModel = CommentId.create(
         new UniqueEntityID(commentId)
       ).getValue();
+
+      const post = await this.postRepo.getPostByPostId(comment.postId);
+      const postResult = post.removeComment(comment);
+      if (postResult.isFailure) {
+        return left(new AppError.UnexpectedError(postResult.getValue()));
+      }
+      await this.postRepo.save(post);
       await this.commentRepo.deleteComment(commentIdModel);
 
       return right(Result.ok<void>());
